[C] Cast to avoid warning converting size_t to int64_t.

+++ aeron-client/src/main/c/concurrent/aeron_term_appender.h
@@ -29,7 +29,7 @@
 inline int64_t aeron_term_appender_get_and_add_raw_tail(volatile int64_t *addr, size_t aligned_length)
 {
     int64_t result;
-    AERON_GET_AND_ADD_INT64(result, *addr, aligned_length);
+    AERON_GET_AND_ADD_INT64(result, *addr, (int64_t)aligned_length);
     return result;
 }
